refactor fonts more
